Added Message Action
this action allows to set a custom message and color on a display
field.
This is particularly useful if you just want to display a text
on a field for example with the pager option.
Or in my case don't display anything for the second page of a
pager if there's no assigned action.